+++ src/cost_functions.rs
@@ -4,36 +4,36 @@ use crate::{common::*, logging};
 // final layer activation and final layer weighted input - SSE uses the former,
 // CrossEntropy will use the latter when I implement it correctly.
 pub trait CostFunction: Copy + Clone {
-    fn cost(&self, final_layer_activation: &Array2<f64>, expected_activation: &Array2<f64>) -> f64;
+    fn cost(&self, final_layer_activation: &Array2<f32>, expected_activation: &Array2<f32>) -> f32;
 
     fn partial_derivative(
         &self,
-        final_layer_activation: &Array2<f64>,
-        expected_activation: &Array2<f64>,
-    ) -> Array2<f64>;
+        final_layer_activation: &Array2<f32>,
+        expected_activation: &Array2<f32>,
+    ) -> Array2<f32>;
 }
 
 #[derive(Copy, Clone)]
 pub struct HalfSSECostFunction;
 
 impl CostFunction for HalfSSECostFunction {
-    fn cost(&self, final_layer_activation: &Array2<f64>, expected_activation: &Array2<f64>) -> f64 {
+    fn cost(&self, final_layer_activation: &Array2<f32>, expected_activation: &Array2<f32>) -> f32 {
         let number_of_examples = column_count(final_layer_activation);
         assert!(final_layer_activation.dim() == expected_activation.dim());
-        (1.0 / number_of_examples as f64)
-            * 0.5_f64
+        (1.0 / number_of_examples as f32)
+            * 0.5_f32
             * expected_activation
                 .iter()
                 .zip(final_layer_activation.iter())
                 .map(|(x, y)| (x - y).powi(2))
-                .sum::<f64>()
+                .sum::<f32>()
     }
 
     fn partial_derivative(
         &self,
-        final_layer_activation: &Array2<f64>,
-        expected_activation: &Array2<f64>,
-    ) -> Array2<f64> {
+        final_layer_activation: &Array2<f32>,
+        expected_activation: &Array2<f32>,
+    ) -> Array2<f32> {
         final_layer_activation - expected_activation
     }
 }
@@ -42,22 +42,22 @@ impl CostFunction for HalfSSECostFunction {
 pub struct CrossEntropyCost;
 
 impl CostFunction for CrossEntropyCost {
-    fn cost(&self, final_layer_activation: &Array2<f64>, expected_activation: &Array2<f64>) -> f64 {
+    fn cost(&self, final_layer_activation: &Array2<f32>, expected_activation: &Array2<f32>) -> f32 {
         let number_of_examples = column_count(final_layer_activation);
         assert!(final_layer_activation.dim() == expected_activation.dim());
-        -(1.0 / number_of_examples as f64)
+        -(1.0 / number_of_examples as f32)
             * final_layer_activation
                 .iter()
                 .zip(expected_activation.iter())
                 .map(|(a, y)| y * a.ln() + (1.0 - y) * (1.0 - a).ln())
-                .sum::<f64>()
+                .sum::<f32>()
     }
 
     fn partial_derivative(
         &self,
-        final_layer_activation: &Array2<f64>,
-        expected_activation: &Array2<f64>,
-    ) -> Array2<f64> {
+        final_layer_activation: &Array2<f32>,
+        expected_activation: &Array2<f32>,
+    ) -> Array2<f32> {
         let mut result = final_layer_activation.clone();
         result
             .iter_mut()
