+++ src/lin_alg/host.rs
@@ -1,22 +1,22 @@
 use crate::lin_alg::Dim;
 use std::ops::{Index, IndexMut};
 
-pub type HostVector = Vec<f64>;
+pub type HostVector = Vec<f32>;
 
 pub struct HostMatrix {
-    raw: Vec<f64>,
+    raw: Vec<f32>,
     dim: Dim,
 }
 
 impl HostMatrix {
     fn zeroes(dim: Dim) -> Self {
-        let raw = vec![0_f64; dim.rows * dim.cols];
+        let raw = vec![0_f32; dim.rows * dim.cols];
         Self { raw, dim }
     }
 }
 
 impl Index<(usize, usize)> for HostMatrix {
-    type Output = f64;
+    type Output = f32;
     fn index(&self, index: (usize, usize)) -> &Self::Output {
         let (i, j) = index;
         &self.raw[i * self.dim.cols + j]
